+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- refactor(sdk): replace installReact with setHostReact for improved plugin integration. (`3acb19dd`)
 - feat(docs): update SDK documentation and improve manifest guidelines. (`a85afcea`)
 
 ## 2.11.1 - 2026-08-03